requirements.txt
requests~=2.32.4
packaging~=24.2

nicegui~=2.20.0
matplotlib~=3.8.4
pillow~=11.2.1
pandas~=2.2.3
httpx~=0.28.1
typing_extensions~=4.13.2
rich~=14.0.0
streamlit~=1.46.1
numpy~=2.2.4
PyYAML~=6.0.2
platformdirs~=4.3.7
scipy~=1.15.2
plotly~=6.2.0
